raisin: disable stubdoms and etherboot in xen
Raisin is about cloning (or downloading) components explicitly and
independently, rather than having the Xen build system clone and build
stuff as part of the build process.

Unfortunately it is not possible to build stubdoms outside the Xen tree
today, so just disable stubdoms for the moment.

Similarly etherboot support requires downloading ipxe, so remove it for
now.

With this patch, building Xen via Raisin doesn't fetch anything from the
web, aside from what is configure in the config file.

Signed-off-by: Stefano Stabellini <stefano.stabellini@eu.citrix.com>